requirements.txt
Django==1.9.12
psycopg2==2.6.1
django-suit==0.2.24
django-selectable==0.9.0
